requirements.txt
Adafruit_DHT==1.4.0
paho-mqtt==1.5.1
